Added initia `eula` mode
